Fixed a number of ugly exception-based crashes in DQClient and DQServer...
This added several bugs currently documented in the Trello with quitting properly and threads misbehaving...

DQClient and DQServer now have a more refined communication and have some features temporarily disabled for performance, this is not regression.

PC has had LastAction added to allow for server-side action processing, an internal API will need to be developed for this implementation.